calculo de rt reajustado

+++ CityGusa-api/src/main/java/com/citygusa/com/citygusaapi/Dto/ControleOperacionalDto.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.time.LocalDate;
 
 @Data
@@ -36,7 +37,7 @@ public class ControleOperacionalDto {
     private Integer acumuladoCarga;
     private Integer acumuladoCargaSeca;
     private BigDecimal mediaHoraCarga;
-    private Integer rt;
+    private BigDecimal rt;
 
 
     public ControleOperacionalDto(ControleOperacionalEntity entity) {
